fixing some bugs in styles
