+++ src/pages/MainPage.jsx
@@ -1,15 +1,15 @@
-
-import { Auth } from "../components/authSection/auth";
-import { Description } from "../components/Description/Description";
+import { Auth } from '../components/MainPage/authSection/auth';
+import { Description } from '../components/Description/Description';
 import ReviewsSlider from '../components/MainPage/ReviewsSlider/ReviewsSlider';
 
-
 const MainPage = () => {
-  return <>
-    <Auth />
-    <Description />
-    <ReviewsSlider />
-  </>
+  return (
+    <>
+      <Auth />
+      <Description />
+      <ReviewsSlider />
+    </>
+  );
 };
 
 export default MainPage;
